+++ lib/flutter_reactive.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_reactive/extensions/state.dart';
 
@@ -62,11 +63,17 @@ class Reactive<T> {
   /// Updates the value and notifies listeners and bound states.
   set value(T newValue) => set(newValue);
 
+  bool _equals(T newValue) {
+    if (value == newValue) return true;
+    if (newValue is List) return listEquals(value as List, newValue as List);
+    return false;
+  }
+
   /// Sets a new value.
   ///
   /// If the value did not change, nothing happens.
   void set(T newValue) {
-    if (_value == newValue && _strict) return;
+    if (_equals(newValue) && _strict) return;
     _value = newValue;
     notify();
   }
@@ -331,5 +338,80 @@ class Reactive<T> {
   @override
   String toString() {
     return _value.toString();
+  }
+}
+
+extension ListReactiveExt<T> on Reactive<List<T>> {
+  /// Transforms the current reactive list into a new reactive list by applying
+  /// optional filtering, sorting, and list operations.
+  ///
+  /// This method listens to changes on the source reactive list and keeps the
+  /// returned reactive list in sync after applying the given transformations.
+  ///
+  /// Parameters:
+  /// - [filter]: A predicate used to filter elements.
+  ///   If null, all elements are kept.
+  /// - [sortBy]: A function that returns a comparable value used to sort elements.
+  ///   If null, no sorting is applied.
+  /// - [sortByDesc]: Whether sorting should be in descending order.
+  ///   Only used when [sortBy] is provided. Defaults to false.
+  /// - [reverse]: Whether to reverse the resulting list.
+  /// - [shuffle]: Whether to shuffle the resulting list randomly.
+  /// - [take]: Limits the number of elements in the resulting list.
+  ///
+  /// Returns:
+  /// A new [Reactive<List<T>>] that reflects the transformed version of the
+  /// source list and updates automatically when the source list changes.
+  ///
+  /// Example:
+  /// ```dart
+  /// final users = Reactive<List<User>>([...]);
+  ///
+  /// final topActiveUsers = users.transform(
+  ///   filter: (u) => u.isActive,
+  ///   sortBy: (u) => u.score,
+  ///   sortByDesc: true,
+  ///   take: 10,
+  /// );
+  /// ```
+  ///
+  /// In this example, the returned reactive list will always contain the
+  /// top 10 active users sorted by score in descending order.
+  Reactive<List<T>> transform({
+    bool Function(T element)? filter,
+    Comparable<dynamic> Function(T element)? sortBy,
+    bool? sortByDesc,
+    bool? reverse,
+    bool? shuffle,
+    int? take,
+  }) {
+    bool Function(T) test = filter ?? (_) => true;
+
+    final r = Reactive(_value.where(test).toList(), _strict);
+
+    listen((l) {
+      var filtered = l.where(test).toList();
+      if (filtered.isNotEmpty) {
+        if (sortBy != null) {
+          filtered.sort((a, b) {
+            if (sortByDesc == true) return sortBy(b).compareTo(sortBy(a));
+            return sortBy(a).compareTo(sortBy(b));
+          });
+        }
+        if (reverse == true) {
+          filtered = filtered.reversed.toList();
+        }
+        if (shuffle == true) {
+          filtered.shuffle();
+        }
+        if (take != null) {
+          filtered = filtered.take(take).toList();
+        }
+
+        r.value = filtered;
+      }
+    });
+
+    return r;
   }
 }